Eliminate compiler warnings

+++ src/execsrc/evio2xml.c
@@ -109,7 +109,8 @@ int main (int argc, char **argv) {
   if((buf==NULL)||(xml==NULL)) {
     int sz=maxbuf*sizeof(unsigned int);
     printf("\n   *** Unable to allocate buffers ***\n\n");
-    printf("\n buf size=%d bytes, addr=0x%x     xml size=%d bytes, addr=0x%x\n\n",sz,buf,sz*EVIO2XML,xml);
+    printf("\n buf size=%d bytes, addr=0x%x     xml size=%d bytes, addr=0x%x\n\n",sz,(unsigned int)buf,
+           sz*EVIO2XML,(unsigned int)xml);
     exit(EXIT_FAILURE);
   }
